add new module `rethnet_rpc_server::filter`
addresses review comment https://github.com/NomicFoundation/hardhat/pull/4197#discussion_r1279948308
