+++ crates/rethnet_rpc_server/src/lib.rs
@@ -1,6 +1,5 @@
 use std::net::{SocketAddr, TcpListener};
 use std::sync::Arc;
-use std::time::{Duration, Instant};
 
 use axum::{
     extract::{Json, State},
@@ -16,7 +15,7 @@ use tracing::{event, Level};
 use rethnet_eth::{
     remote::{
         client::{Request as RpcRequest, RpcClient},
-        filter::{FilterBlockTarget, FilterOptions, FilteredEvents, LogOutput, OneOrMoreAddresses},
+        filter::{FilteredEvents, LogOutput},
         jsonrpc,
         jsonrpc::{Response, ResponseData},
         methods::MethodInvocation as EthMethodInvocation,
@@ -38,6 +37,9 @@ pub use hardhat_methods::{
 
 mod config;
 pub use config::{AccountConfig, Config};
+
+mod filter;
+use filter::{new_filter_deadline, Filter};
 
 #[derive(Clone, Copy)]
 struct U256WithoutLeadingZeroes(U256);
@@ -63,72 +65,6 @@ pub enum MethodInvocation {
 }
 
 type RethnetStateType = Arc<RwLock<dyn SyncState<StateError>>>;
-
-struct _FilterCriteria {
-    _from_block: U256,
-    _to_block: U256,
-    _addresses: Vec<Address>,
-    _topics: Vec<B256>,
-}
-
-impl _FilterCriteria {
-    async fn _from_request_and_state<T>(
-        request_options: FilterOptions,
-        state: StateType,
-    ) -> Result<Self, ResponseData<T>> {
-        let (_from_block, _to_block) = match request_options.block_target {
-            Some(FilterBlockTarget::Hash(hash)) => {
-                let block_number = _block_number_from_hash(&state, &hash, false)?;
-                (block_number, block_number)
-            }
-            Some(FilterBlockTarget::Range { from, to }) => {
-                let from =
-                    _block_number_from_block_spec(&state, &from.unwrap_or(BlockSpec::latest()))
-                        .await?;
-                let to = match to {
-                    None => from,
-                    Some(to) => _block_number_from_block_spec(&state, &to).await?,
-                };
-                (from, to)
-            }
-            None => {
-                let latest_block_number = get_latest_block_number(&state).await?;
-                (latest_block_number, latest_block_number)
-            }
-        };
-
-        let _addresses = match request_options.addresses {
-            Some(OneOrMoreAddresses::One(address)) => vec![address],
-            Some(OneOrMoreAddresses::Many(addresses)) => addresses,
-            None => Vec::new(),
-        };
-
-        let _topics = request_options.topics.unwrap_or(Vec::new());
-
-        Ok(Self {
-            _from_block,
-            _to_block,
-            _addresses,
-            _topics,
-        })
-    }
-}
-
-struct Filter {
-    // TODO: later, when adding in the rest of the filter methods, consider removing this `type`
-    // field entirely.  i suspect it's probably redundant as compared to the type implied by the
-    // variants in `events`, but that suspicion will be confirmed or denied by the addition of
-    // those other filter methods.
-    // r#type: SubscriptionType,
-    _criteria: Option<_FilterCriteria>,
-    deadline: std::time::Instant,
-    events: FilteredEvents,
-    is_subscription: bool,
-}
-
-fn new_filter_deadline() -> Instant {
-    Instant::now() + Duration::from_secs(5 * 60)
-}
 
 struct AppState {
     rethnet_state: RethnetStateType,
